Moved SwitchSelector to component folder
