+++ Car/src/components/SwitchSelector/SwitchSelectorProps.ts
