+++ Car/src/components/SwitchSelector/SwitchSelectorStyle.ts
